Refactored the MentorEdit component
